requirements.txt
asgiref==3.6.0
Django==4.2
django-environ==0.10.0
django-rest-framework==0.1.0
djangorestframework==3.14.0
python-environ==0.4.54
pytz==2023.3
sqlparse==0.4.3